# swagger.yaml
openapi: 3.0.0
info:
  description: "This is the swagger documentation and API for the GeoCode project implemented by Peak Performers for the client [5DT](https://5dt.com/)."
  version: "1.0.0"
  title: "Swagger GeoCode"
  termsOfService: "https://swagger.io/terms/"
  contact:
    email: "peakperformers@geocodeapp.tech"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: 'http://geocodeapp.tech:8080/api'
...
paths:
...
  /Collectable/createCollectableType:
    post:
      tags:
        - "Collectable"
      summary: "Creates a new Collectable Type"
      description: "Create Collectable Type"
      operationId: "createCollectableType"
      requestBody:
        required: true
        description: "Request to create a new Collectable Type"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectableTypeRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateCollectableTypeRequest'
      responses:
        '200':
          description: "Collectable Type created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectableType'
            application/xml:
              schema:
                $ref: '#/components/schemas/CollectableType'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/createCollectable:
    post:
      tags:
        - "Collectable"
      summary: "Creates a new Collectable"
      description: "Create Collectable"
      operationId: "createCollectable"
      requestBody:
        required: true
        description: "Request to create a new Collectable"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectableRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateCollectableRequest'
      responses:
        '200':
          description: "Collectable created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collectable'
            application/xml:
              schema:
                $ref: '#/components/schemas/Collectable'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/createCollectableSet:
    post:
      tags:
        - "Collectable"
      summary: "Create a new Collectable Set"
      description: "Create Collectable Set"
      operationId: "createCollectableSet"
      requestBody:
        required: true
        description: "Request to create a new Collectable Set"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectableSetRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateCollectableSetRequest'
      responses:
        '200':
          description: "Successfully created Collectable Set"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectableSet'
            application/xml:
              schema:
                $ref: '#/components/schemas/CollectableSet'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/getCollectables:
    get:
      tags:
        - "Collectable"
      description: "Returns all Collectables"
      operationId: "getCollectables"
      responses:
        '200':
          description: "Successfully returned Collectables"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectablesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectablesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/getCollectableTypes:
    get:
      tags:
        - "Collectable"
      description: "Returns all Collectable Types"
      operationId: "getCollectableTypes"
      responses:
        '200':
          description: "Successfully returned Collectable Types"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectableTypesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectableTypesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/getCollectableSets:
    get:
      tags:
        - "Collectable"
      description: "Returns all Collectable Sets"
      operationId: "getCollectableSets"
      responses:
        '200':
          description: "Successfully returned Collectable Sets"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectableSetsResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectableSetsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/getCollectableByType:
    post:
      tags:
        - "Collectable"
      description: "Returns all Collectables of the specified type"
      operationId: "getCollectableByType"
      requestBody:
        required: true
        description: "Request Collectables by type"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCollectableByTypeRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetCollectableByTypeRequest'
      responses:
        '200':
          description: "Successfully returned Collectables of given type"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectablesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectablesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/getCollectableTypeByRarity:
    post:
      tags:
        - "Collectable"
      description: "Returns all Collectable Types of the specified rarity"
      operationId: "getCollectableTypeByRarity"
      requestBody:
        required: true
        description: "Request Collectable Types by rarity"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCollectableTypeByRarityRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetCollectableTypeByRarityRequest'
      responses:
        '200':
          description: "Successfully returned Collectable Types of given rarity"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectableTypesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectableTypesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /Collectable/getCollectableTypeBySet:
    post:
      tags:
        - "Collectable"
      description: "Returns all Collectable Types of the specified set"
      operationId: "getCollectableTypeBySet"
      requestBody:
        required: true
        description: "Request Collectable Types by set"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCollectableTypeBySetRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetCollectableTypeBySetRequest'
      responses:
        '200':
          description: "Successfully returned Collectable Types of given set"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectableTypesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectableTypesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
components:
  responses:
    UnauthorizedError:
      description: Invalid JWT token
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Rarity:
      type: string
      enum:
        - "COMMON"
        - "UNCOMMON"
        - "RARE"
        - "EPIC"
        - "LEGENDARY"
...
    CollectableType:
      type: "object"
      required:
        - id
        - name
        - image
        - rarity
        - set
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        image:
          type: string
        rarity:
          $ref: "#/components/schemas/Rarity"
        set:
          $ref: "#/components/schemas/CollectableSet"
      xml:
        name: "CollectableType"
    CollectableSet:
      type: "object"
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
      xml:
        name: "CollectableSet"
    Collectable:
      type: "object"
      required:
        - id
        - type
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/CollectableType"
      xml:
        name: "Collectable"
    CreateCollectableSetRequest:
      type: "object"
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
      xml:
        name: "CreateCollectableSetRequest"
    CreateCollectableTypeRequest:
      type: "object"
      required:
        - name
        - image
        - rarity
        - setId
      properties:
        name:
          type: string
        image:
          type: string
        rarity:
          $ref: "#/components/schemas/Rarity"
        setId:
          type: string
          format: uuid
      xml:
        name: "CreateCollectableTypeRequest"
    CreateCollectableRequest:
      type: "object"
      required:
        - collectableTypeId
      properties:
        collectableTypeId:
          type: string
          format: uuid
      xml:
        name: "CreateCollectableRequest"
    GetCollectablesResponse:
      type: "object"
      required:
        - collectables
      properties:
        collectables:
          type: array
          items:
            $ref: '#/components/schemas/Collectable'
      xml:
        name: "CollectableResponse"
    GetCollectableTypesResponse:
      type: "object"
      required:
        - collectableTypes
      properties:
        collectableTypes:
          type: array
          items:
            $ref: '#/components/schemas/CollectableType'
      xml:
        name: "CollectableTypeResponse"
    GetCollectableSetsResponse:
      type: "object"
      required:
        - collectableSets
      properties:
        collectableSets:
          type: array
          items:
            $ref: '#/components/schemas/CollectableSet'
    GetCollectableByTypeRequest:
      type: "object"
      required:
        - collectableTypeId
      properties:
        collectableTypeId:
          type: string
          format: uuid
      xml:
        name: "GetCollectableByTypeRequest"
    GetCollectableTypeByRarityRequest:
      type: "object"
      required:
        - rarity
      properties:
        rarity:
          $ref: "#/components/schemas/Rarity"
      xml:
        name: "GetCollectableTypeByRarity"
    GetCollectableTypeBySetRequest:
      type: "object"
      required:
        - setId
      properties:
        setId:
          type: string
          format: uuid
      xml:
        name: "GetCollectableTypeBySet"
...